feat:update clean.sh

+++ clean.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
+###################### docker ######################
 HOME="/Users/zhaohongliang/DockerData"
-
 NETWORK="canary-net"
 
+###################### mysql #######################
 MYSQL_HOME="$HOME/mysql"
 
+##################### proxysql ######################
 PROXYSQL_HOME="$HOME/proxysql"
 
 # 定义旋转的光标符号
@@ -18,32 +20,52 @@ BLUE='\033[0;34m'   # 蓝色字体
 BOLD='\033[1m'      # 加粗字体
 NC='\033[0m'        # 恢复默认格式
 
-# 停止服务并清理
-function clean() {
-    while [ $(docker ps -aq --filter "name=sql" | wc -l) -gt 0 ]; do
-        for i in $(seq 0 3); do
-            printf "\rStopping \`mysql\`...${SPINNER:$i:1}"   # 使用回车符（\r）实现光标回到行首
-            sleep 0.2
-        done
-        docker stop `docker ps -a | grep sql | awk -F " " '{print $1}'` &> /dev/null
-        docker rm `docker ps -a | grep sql | awk -F " " '{print $1}'` &> /dev/null
-        docker network rm $NETWORK &> /dev/null
-    done
+# 移除容器
+function removeContainers() {
+    container_number=$(docker ps -aq --filter "name=sql" | wc -l)
+    container_filter=$(docker ps -a | grep sql | awk -F " " '{print $1}')
+    if [ $container_number -gt 0 ]; then
+        echo "Stopping mysql and proxysql...(might take a while)"
+        docker stop $container_filter &> /dev/null
+        sleep 2
+        echo "${BLUE}===>${NC}${BOLD} Successfully stopped mysql and proxysql${NC}"
+        
+        docker rm $container_filter &> /dev/null
+        sleep 2
+        echo "${BLUE}===>${NC}${BOLD} Successfully removed mysql and proxysql${NC}"
+    else
+        echo "MySQL and ProxySQL containers stopped and removed."
+    fi
+}
 
-    printf "\r\033[K"
-    printf "Stopping \`mysql\`..."
-    printf "\n${BLUE}===>${NC}${BOLD} Successfully stopped \`mysql\`!${NC}\n"
-    
-    while true; do
-        read -p "Do you want to keep the data and configuration files?（y/n）" choice
-        choice_lower=$(printf "$choice" | tr '[:upper:]' '[:lower:]')
-        if [ "$choice_lower" == "y"  ] || [ "$choice_lower" == "yes"  ]; then
-            break
-        elif [ "$choice_lower" == "n"  ] || [ "$choice_lower" == "no"  ]; then
-            rm -rf $MYSQL_HOME && rm -rf $PROXYSQL_HOME
-            break
+# 移除网路
+function removeNetwork() {
+    flag=0
+    while [ $flag -ne 1 ]; do
+        read -p "Do you want to keep the $NETWORK network?（Y/N）" choice
+        if [ "$choice" == "Y"  ] || [ "$choice" == "y" ]; then
+            flag=1
+        elif [ "$choice" == "N"  ] || [ "$choice" == "n" ]; then
+            docker network rm $NETWORK &> /dev/null
+            flag=1
         else
-            printf "Invalid choice. Please enter 'yes' or 'no'\n"
+            echo "Invalid choice. Please enter 'Y' or 'N'"
+        fi
+    done
+}
+
+# 删除配置文件和数据
+function removeData() {
+    flag=0
+    while [ $flag -ne 1 ]; do
+        read -p "Do you want to keep the data and configuration files?（Y/N）" choice
+        if [ "$choice" == "Y"  ] || [ "$choice" == "y" ]; then
+            flag=1
+        elif [ "$choice" == "N"  ] || [ "$choice" == "n" ]; then
+            rm -rf $MYSQL_HOME && rm -rf $PROXYSQL_HOME
+            flag=1
+        else
+            echo "Invalid choice. Please enter 'Y' or 'N'"
         fi
     done
 }
@@ -68,7 +90,9 @@ function logo()
 
 # main
 function main() {
-    clean
+    removeContainers
+    removeNetwork
+    removeData
     logo
 }
